sizes

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("Size of a char: %lu byte(s)", sizeof(a));
-	printf("Size of an int: %lu byte(s)", sizeof(b));
-	printf("Size of a long int: %lu byte(s)", sizeof(c));
-	printf("Size of a long long int: %lu byte(s)", sizeof(d));
-	printf("Size of a float: %lu byte(s)", sizeof(e));
+	printf("Size of a char: %lu byte(s)", (unsigned long)sizeof(a));
+	printf("Size of an int: %lu byte(s)", (unsigned long)sizeof(b));
+	printf("Size of a long int: %lu byte(s)", (unsigned long)sizeof(c));
+	printf("Size of a long long int: %lu byte(s)", (unsigned long)sizeof(d));
+	printf("Size of a float: %lu byte(s)", (unsigned long)sizeof(e));
 	return (0);
 }
